+++ docs/changelog.rst
@@ -23,6 +23,7 @@ v1.14.0 - 2018-xx-xx
 
 **Tickets resolved in this release:**
 
+- [`#441 <https://github.com/dennissiemensma/dsmr-reader/issues/441>`_] PVOutput exports schedulen naar ingestelde upload interval - by pyrocumulus
 - [`#436 <https://github.com/dennissiemensma/dsmr-reader/issues/436>`_] Provide built-in authentication method for public webinterface